requirements.txt
Django>=5.0,<6.0
openpyxl>=3.1
tzdata>=2024.1
Pillow>=10.0
psycopg2-binary>=2.9
waitress>=3.0
pytest>=8.2
pytest-django>=4.8
